Fixes #32534 - Keep "old core" tests only
