+++ Rakefile
@@ -11,17 +11,8 @@ namespace :test do
     ENV['DYNFLOW_DB_CONN_STRING'] = 'sqlite:/'
     t.libs << '.'
     t.libs << 'lib'
-    t.libs << 'test/core_test'
-    t.test_files = FileList['test/core_test/*_test.rb']
-  end
-
-  desc 'Test Dynflow api plugin'
-  Rake::TestTask.new(:api) do |t|
-    ENV['DYNFLOW_DB_CONN_STRING'] = 'sqlite:/'
-    t.libs << '.'
-    t.libs << 'lib'
-    t.libs << 'test/api_test'
-    t.test_files = FileList['test/api_test/*_test.rb']
+    t.libs << 'test'
+    t.test_files = FileList['test/*_test.rb']
   end
 end
