--- test/core_test/test_helper.rb
@@ -1,14 +1,0 @@
-require 'minitest/autorun'
-
-ENV['RACK_ENV'] = 'test'
-
-$LOAD_PATH << File.join(File.dirname(__FILE__), '..', '..', 'lib')
-require "mocha/setup"
-require "rack/test"
-
-require 'dynflow'
-require 'smart_proxy_dynflow_core'
-require 'smart_proxy_dynflow_core/testing'
-
-WORLD = SmartProxyDynflowCore::Dynflow::Testing.create_world
-SmartProxyDynflowCore::Core.instance.world = WORLD
