+++ test/task_launcher_registry_test.rb
